bypass check_hero() and add 2 similar functions
Here have been added 3 funcs which check the heros state.
Two of these are only used in this segment, so they are not mentioned
in the header.

Also its interesting what bit 5 of the heros state means and
whats up if *(hero+0x84) == 0x10.
That disables the hero somewhat. Maybe these states are temporary,
e.g. in fights in combinations with timers.