bypass decomp_rle()
THE INCREDIBLE PERFORMANCE BOOST, till now.
This is the decompressor for the RLE pictures, used for the textures
in cities and dungeons. It is called very often at each step or direction
change.

With DOSBox-v0.73, core=normal, cycles=2000, frameskip=0 and output=surface
I get a fluid walk feeling on my
AMD Athlon(tm) 64 X2 Dual Core Processor 4600+ with 2400 MHz.