initial bypassing of nearcalls for simple/normal
Now we have support for bypassing nearcalls for the cores simple and normal.

This feature should be handled with care, since near calls
are trickier than far calls.

Nearcall bypassing in DSA1 should only be used temporary for two reasons.

* The dynamic core does (atm) not support it, but should make no problems.

* If all callers in a segment are bypassed, the cather for the near function
	should be removed to reduce overhead.

There are different nearcall scenarios, so beware!
In DSA1 we have often far functions called as near functions
with a CS pushed to the stack before the call.

	push cs
	call near func_with_retf

If we want a bypass, we have to remove the return address (4 bytes CS:IP)
from the stack.

But there are functions, which are true nearcalls.
To bypass them, we have to remove the return address also,
which are 2 bytes (IP).

And beware of RETF and RET with an immediate value, for stack resetting.