rework pic_copy in mode 1
The loop here is at least executed one time, so a do {} while is
the better way.
Also there should be  check to avoid overwriting pixel in some regions with
the colors 0xc8-0xff.
This is now added and shows the city- and dungeon maps without the compass,
like in the original.