uncomment to use EMS (buggy!)
The game was intentionally developed with EMS support, to avoid
slow disk access.

Tragically it was disabled. Luckily this was done by a variable
and Borland C++ was not able to remove that dead code.

I found two Bugs with that:

Re-entering the map screen, leaves the map black.

In fight mode sometime huge sizes of memory are tried to be allocated.

If you are brave, you can remove the comment.