Original-Bugfix: infinite-loop in get_random_hero()
The function get_random_hero() can loop forever,
if all positions of the current group heros are
greater than the number of heros in the group.

It can best figured out with one hero who is not in the first position.
If you walk through the city and trigger
a city event (1% in Thorwal) you often get this endless loop.

For details see:
https://github.com/Henne/Bright-Eyes/issues/closed#issue/5