add a timer replacement
Every 1/18.2 seconds a timer interrupt is done on a DOS-PC.
This interrupts the running program an executes a timer interrupt routine.
Since we now have modern Operating systems, we don't have access to the timer,
but SDL provides a timer subsystem which uses threads.
The difference is, that our main program (schickm.exe in our case) does not
stop when the timer occours, which can lead to race conditions.

At the moment we use this timer only to change the variable seed2,
which is needed to produce random numbers.