G105de: Bugfix: wrong mage spells in novice mode
In novice mode, where you don't increase the spells manually,
an array[45] is generated, which content depends on the mages school.
The content are the spells, which shoud be tried
to be increased automatically.

And there are house spells, which may be a subset of the school spells.

The values in this array are build as:
1. all house spells,
2. all school spells,
3. 5 constant domination spells
4. all house spells,
5. all house spells,
6. the rest is filled up with random spells

The bug was, that I accidently filled 4. and 5. with school spells,
instead of house spells.