M302de: BC: add other Prototypes
The BC does seems to be a bit lumpy.
The arguments of functions in c files
do not have to be in sync with that one in the headers.

Here in these function the argument pos is long,
but other functions call it with short.
If the function argument would be short, it would have been
copied to a register at the start of the function.
That would result in different code.