M302de: Original-Bugfix: some item counter fixes
The item counter in the character sheet gets broken when items
get moved in the inventory. This is bad, since that item counter
is referenced in SCHWEIF when importing from SCHICK, and crahes
when the value is too high (or too low).

This bug is known for some time and Bright-Eyes controlles the
item counter with a timer and corrects it.

Now two places have been fixed, when two stackable items of the same
type are merged.