M302de: make output of Unknown timers more concrete
These two unknown timers in the character sheet are now known.
One timer is the time until the next heal attempt on this character
can be done.
The other timer is the time until the next attempt to increase the
spellstaff level.

They now produce only outputs, when they reach zero,
which may be useful to the gamer.