M302de: Original-Bugfix: dont confuse escape fields with heros
If a hero searches a target for a spell,
it can happen that the search reaches a escape field,
which has the ID = -1.

This will be interpreted by the original code as a hero,
and the game will check if this hero is alive and well.

On newer systems this produces, of course,  a SEGFAULT.