M302de: Original-Bugfix: lantern refill may not work
A burning lantern can be refilled only by having a flask of oil.
But the program looks, for the oil and the burning lantern,
in the inventory of the hero which cast a spell most recently.

This might be the hero who had enlighten the lantern,
since this is implemented by using a spell and temporarily
make this hero a spellcaster, even if he (or she) is not
able to do magic.

Reproduce:
Prequisites:
* 2 Heros: 1 with magic abilities (Hero A) and another (Hero B)
* Hero B: has Oil (2x), lantern (1x, not burning), thinder (1x)
* Hero A: has no oil

* Hero B: use thinder => enlightens the lantern
* Hero A: cast a healing spell on Hero B => program remembers Hero A as the SPELLCASTER
* Hero B: use the burning lantern to refill => "ERROR: Hero B has no oil"