M302de: BC: correct prototypes of seg044
Functions in C/CPP files should match exactly to their counterparts in the
header files.

Some C/CPP files do not include their own header and may get out of sync.

This one could be fixed by:

a) use the types in the header, which make calling functions have the same code
b) casting each reference of the argument in the called function

Sad, but true. :(