M302de: Bugfix: spell animations at the wrong place
When casting, for example, "Saft Kraft" the animation of
the spell was at the wrong place.

The code, which BCC compiled to an identical binary,
does not work in the same way on GCC and MSVC.
And it was ugly, too.

Heres a not so ugly and more portable version of it.