M302de: Original-Bugfix: riddle-chest on the ship of death
If the user answers "MARBO", the correct answer of the question,
a string is generated from the file SHIP.DTX entry 14.
This string expects also the name of the group leader,
bu accidently the programmers had given the integer 10,
which seems to be the AP each groupmember should get.

The game crashes on modern architectures.