M302de: Bugfix: disable bypass of schick_main()
The JoWood-release of SCHICK had no Audio-CD data.
A patch of me iin SCHICKM.EXE enabled MIDI-sound for those people.

But this patch will not work anymore when the bypass of schick_main() is
active until I enable MIDI playback in schick_main().

This needs some time, to think, so this bypass gets disabled.

Thanks to uxl for remembering:
http://www.crystals-dsa-foren.de/showthread.php?tid=700&pid=150999#pid150999